# swagger.yml
openapi: 3.0.1
info:
  title: ThinkLab Manager API
  description: REST API for managing system users and related operations.
  contact:
    name: Development Team
    email: dev@thinklab.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: "0.1"
...
paths:
  /assets:
    get:
      tags:
      - Asset Management
      summary: Search or list assets
      description: Retrieves a list of assets filtered optionally by physical asset
        tag or serial number.
      operationId: listOrSearchAssets
      parameters:
      - name: assetTag
        in: query
        description: Filter by exact corporate asset tag
        explode: false
        schema:
          type: string
          nullable: true
      - name: serialNumber
        in: query
        description: Filter by exact hardware serial number
        explode: false
        schema:
          type: string
          nullable: true
      responses:
        "200":
          description: Query executed successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssetManagementResponse"
    post:
      tags:
      - Asset Management
      summary: Register a new asset
      description: Creates a new hardware asset and initializes its centralized audit
        trail.
      operationId: createAsset
      parameters:
      - name: X-Executor
        in: header
        description: Identifier of the entity executing the creation
        required: true
        schema:
          minLength: 1
          type: string
      requestBody:
        description: The validated JSON payload containing asset specifications.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetManagementRequest"
        required: true
      responses:
        "201":
          description: Asset successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetManagementResponse"
        "400":
          description: Invalid request payload
        "409":
          description: Asset tag or serial number already exists
  /assets/{assetHash}:
    get:
      tags:
      - Asset Management
      summary: Retrieve an asset
      description: Fetches the current specifications and lifecycle state of an asset
        by its hash.
      operationId: getAsset
      parameters:
      - name: assetHash
        in: path
        description: The unique business hash of the asset
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Asset found and returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetManagementResponse"
        "404":
          description: Asset not found
    put:
      tags:
      - Asset Management
      summary: Update an asset
      description: Performs a full or partial update on an asset and logs granular
        from/to changes in the audit trail.
      operationId: updateAsset
      parameters:
      - name: assetHash
        in: path
        description: The unique business hash of the asset
        required: true
        schema:
          type: string
      - name: X-Executor
        in: header
        description: Identifier of the entity executing the update
        required: true
        schema:
          minLength: 1
          type: string
      requestBody:
        description: The validated JSON payload containing the new values.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetManagementRequest"
        required: true
      responses:
        "200":
          description: Asset successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetManagementResponse"
        "400":
          description: Invalid request payload
        "404":
          description: Asset not found
        "409":
          description: Asset tag or serial number conflict
    delete:
      tags:
      - Asset Management
      summary: Delete an asset
      description: Permanently removes the hardware asset record from the inventory
        database.
      operationId: deleteAsset
      parameters:
      - name: assetHash
        in: path
        description: The unique business hash of the asset
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Asset successfully deleted
        "404":
          description: Asset not found
...
components:
  schemas:
...
    AssetManagementRequest:
      required:
      - assetTag
      - equipmentType
      - purchaseValue
      - serialNumber
      - specification
      type: object
      properties:
        assetTag:
          minLength: 1
          type: string
          x-size-message: Asset tag cannot be blank
        serialNumber:
          minLength: 1
          type: string
          x-size-message: Serial number cannot be blank
        equipmentType:
          $ref: "#/components/schemas/EquipmentType"
        status:
          nullable: true
          allOf:
          - $ref: "#/components/schemas/AssetStatus"
        userId:
          type: string
          nullable: true
        purchaseValue:
          minimum: 0
          type: number
          x-minimum-message: Purchase value must be greater than or equal to zero
          x-not-null-message: Purchase value is required
        saleValue:
          minimum: 0
          type: number
          nullable: true
          x-minimum-message: Sale value must be greater than or equal to zero
        acquisitionDate:
          type: string
          format: date-time
          nullable: true
        specification:
          $ref: "#/components/schemas/AssetSpecificationModel"
      description: Data Transfer Object representing the incoming payload to create
        or update a hardware asset. Contains rigorous Jakarta validation rules to
        enforce data compliance at the API edge.
    AssetManagementResponse:
      required:
      - acquisitionDate
      - assetHash
      - assetTag
      - comments
      - equipmentType
      - purchaseValue
      - serialNumber
      - specification
      - status
      type: object
      properties:
        assetHash:
          type: string
        assetTag:
          type: string
        serialNumber:
          type: string
        equipmentType:
          $ref: "#/components/schemas/EquipmentType"
        status:
          $ref: "#/components/schemas/AssetStatus"
        userId:
          type: string
          nullable: true
        purchaseValue:
          type: number
        saleValue:
          type: number
          nullable: true
        acquisitionDate:
          type: string
          format: date-time
        decommissionDate:
          type: string
          format: date-time
          nullable: true
        saleDate:
          type: string
          format: date-time
          nullable: true
        specification:
          $ref: "#/components/schemas/AssetSpecificationModel"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/MaintenanceComment"
      description: Data Transfer Object representing the unified outbound payload
        for asset profiles. Formatted for seamless integration with client-side user
        interfaces and frontend dashboards.
    AssetSpecificationModel:
      type: object
      properties:
        manufacturer:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        macAddress:
          type: string
          nullable: true
        ipAddress:
          type: string
          nullable: true
        operatingSystem:
          type: string
          nullable: true
      description: Embedded record representing structural hardware specifications
        and networking identities. Directly nested within the asset document inside
        MongoDB.
    AssetStatus:
      type: string
      description: Enumeration representing the granular administrative lifecycle
        states of a corporate hardware asset.
      enum:
      - AVAILABLE
      - ALLOCATED
      - UNDER_MAINTENANCE
      - DECOMMISSIONED
      - SOLD
      x-enum-descriptions:
      - The asset is sitting in inventory and is completely ready to be assigned to
        an employee.
      - The asset has been actively provisioned and is currently in custody of an
        employee.
      - "The device is undergoing physical repairs, hardware upgrades, or configuration\
        \ resets."
      - "The hardware has reached its logical end-of-life, is broken, or was retired\
        \ from active service."
      - The asset was legally sold or liquidated to a third party or employee.
...
    EquipmentType:
      type: string
      description: Enumeration detailing the structural taxonomy classifications of
        corporate hardware items.
      enum:
      - LAPTOP
      - DESKTOP
      - MOBILE
      - NETWORKING
      - SERVER
      x-enum-descriptions:
      - "Portable employee workstations (e.g., MacBooks, ThinkPads, Dell Latitudes)."
      - Fixed desk workstations or mini PC hosts.
      - Corporate mobile cellular nodes or tablets.
      - "Network infrastructure units (e.g., switches, physical routers, firewalls,\
        \ Wi-Fi APs)."
      - On-premise infrastructure hosts or rack-mounted computational blocks.
...
    MaintenanceComment:
      required:
      - author
      - message
      - statusWhen
      - timestamp
      type: object
      properties:
        author:
          type: string
          description: The name or business identifier of the person who wrote the
            comment.
        message:
          type: string
          description: The actual discursive content/text of the update.
        statusWhen:
          $ref: "#/components/schemas/MaintenanceStatus"
        timestamp:
          type: string
          description: The absolute instant when the comment was recorded.
          format: date-time
      description: Represents an immutable user or technician comment within a maintenance
        ticket. Captures historical timeline discourse bound to the system state at
        that specific moment.
    MaintenanceStatus:
      type: string
      description: "Defines the strict, chronological lifecycle states for a computer\
        \ maintenance ticket. Annotated with Serdeable to allow seamless compilation\
        \ and native JSON serialization."
      enum:
      - OPENED
      - IN_ANALYSIS
      - AWAITING_PARTS
      - EXTERNAL_VENDOR
      - MANUFACTURER_WARRANTY
      - AWAITING_PICKUP
      - TESTING
      - COMPLETED
      - CANCELED
      x-enum-descriptions:
      - Ticket successfully registered and awaiting initial technical triage.
      - Computer is currently on the technical bench undergoing hardware diagnostics.
      - "Diagnosis completed, but progress is halted waiting for internal or external\
        \ parts replacement."
      - Ticket paused because the equipment was dispatched to a specialized external
        vendor.
      - Ticket paused because the equipment was dispatched to the manufacturer's official
        warranty service.
      - Maintenance or external execution finished; awaiting pickup from vendor or
        notification to user.
      - "Equipment returned to bench for hardware stress tests, quality assurance,\
        \ and OS stabilization."
      - "Maintenance flow successfully validated, resolved, and equipment returned\
        \ to the user."
      - "Ticket administrative cancellation due to irreparable damage, structural\
        \ write-off, or duplication."
...